Long time no C
It's been a while since I've worked on the project.
Made the application to be a server (and I might need to create a seperate application similar to this one, only that it will be a client). Also, I might have to rework a big chunk of code: I'll remove the LocatieTabla.Muta method and make a very similar function defined in Form1