Communication
I've done it. Made 2 seperate applications: the server(this application) and the client. Everytime a move occurs, the coordinates are sent to the other app, where that move will be executed as well. Also, there's a chatbox for these two players.
Just a few last touches and the project should be done.