Final form
At least for the faculty project.
It's been a great and learning experience. By doing this project I've learned a lot about POO, but I've also learned that what I know so far is just the surface of what programming really is. There is much to be discoverd and to be tinkered with, and this project proved that point.
It's been a pleasure doing this project :)